Fix test being silently ignored.
There were two classes TestEMBLFeature meaning the first one and its
single test were being ignored.

The test also needed updated to wrap at 80 rather than 79 characters
(after recent commits), and not expect a trailing newline.

However, the test itself was meant to verify #50 was fixed, but didn't.
The test needs changing as we do not want the long word to be split.